Create a randomly moving target.
This commits adds code for a randomly moving target.  In the process of making
this addition, I also added several new settings and reorganized some of the
GUI code.

I also stopped the sight from being updated, rendered, or even created.  All
of these changes would have to be undone in order to keep working on the sight
code.