Add network support.
This changed involved a few architectural changes.  Most prominently, I added
the concept of a "system" to the main script.  Systems are just objects that
are setup before the game loop, updated every frame during the game, and torn
down afterwards.  Most of this code was already there, I just formalized it a
bit.