Debug substantial sections of the post office.
At this point, the post office seems to work more or less.  There could be a
few bugs left, but I don't think they will be hard to find or fix.

The biggest remaining problem has to do with how tokens are constructed.
Presently, they all contain a reference to the world.  This is problematic,
because the world can't be pickled and sent over the network.  I need to come
up with a new way to construct the game objects that is more conducive to
pickling.